WIP refactor code for better separation of concerns between parsing and export. Also remove plotting functionality which bloats this library
